change code structure based on code review: split ActionButton to multiple components
