+++ components/CreateButton/index.js
@@ -1,0 +1,12 @@
+import { Icon } from "@iconify/react";
+import { StyledActionButton } from "../StyledActionButton/ActionButton.styled";
+
+function CreateButton() {
+  return (
+    <StyledActionButton href="/form">
+      <Icon icon="ic:round-add" width={32} aria-label="Add new expense" />
+    </StyledActionButton>
+  );
+}
+
+export default CreateButton;
