+++ components/GoBackButton/index.js
@@ -1,0 +1,15 @@
+import { useRouter } from "next/router";
+import { StyledActionButton } from "../StyledActionButton/ActionButton.styled";
+import { Icon } from "@iconify/react";
+
+function GoBackButton() {
+  const router = useRouter();
+
+  return (
+    <StyledActionButton as="button" onClick={() => router.back()}>
+      <Icon icon="ic:sharp-arrow-back" width={32} aria-label="Go back" />
+    </StyledActionButton>
+  );
+}
+
+export default GoBackButton;
